moves createnewteamdialog to its own file
